pyproject.toml
[project]
name = "openaiagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiodns==3.5.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "black>=25.1.0",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "distro==1.9.0",
    "fastapi==0.116.2",
    "flake8>=7.3.0",
    "frozenlist==1.7.0",
    "griffe==1.14.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "jiter==0.10.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "mcp==1.13.1",
    "multidict==6.6.4",
    "openai==1.106.1",
    "openai-agents==0.2.11",
    "propcache==0.3.2",
    "pycares==4.11.0",
    "pycparser==2.23",
    "pydantic[email]==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart==0.0.20",
    "passlib[bcrypt]>=1.7.4",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rpds-py==0.27.1",
    "sniffio==1.3.1",
    "sse-starlette==3.0.2",
    "starlette==0.47.3",
    "tqdm==4.67.1",
    "types-requests==2.32.4.20250809",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "vulture>=2.14",
    "websockets>=15.0.1",
    "yarl==1.20.1",
    "perplexityai>=0.17.1",
]
